feat: protege as rotas

+++ favorite_product_api/users/routes.py
@@ -4,6 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, Response, status
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from favorite_product_api.auth.auth import get_current_user
 from favorite_product_api.databases.postgresql import get_db
 from favorite_product_api.exceptions import NotFoundError
 from favorite_product_api.users.payloads import CreateUserPayload, UpdateUserPayload
@@ -23,7 +24,10 @@ async def create_user(
 ):
     try:
         user = await UserService.create_user(
-            db_session=db_session, name=payload.name, email=payload.email
+            db_session=db_session,
+            name=payload.name,
+            email=payload.email,
+            password=payload.password,
         )
         return CreateUserResponse(uuid=user.uuid, email=user.email, name=user.name)
     except ValueError as e:
@@ -46,8 +50,16 @@ async def create_user(
     "/{uuid}", response_model=UpdateUserResponse, response_model_exclude_none=True
 )
 async def update_user(
-    uuid: UUID, payload: UpdateUserPayload, db_session: AsyncSession = Depends(get_db)
+    uuid: UUID,
+    payload: UpdateUserPayload,
+    db_session: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user),
 ):
+    if current_user.uuid != uuid:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Operation not permitted",
+        )
     try:
         updated_data = await UserService.update_user(
             db_session=db_session, uuid=uuid, name=payload.name, email=payload.email
@@ -72,7 +84,11 @@ async def update_user(
 
 
 @router.get("/{uuid}", response_model=GetUserResponse)
-async def get_user(uuid: UUID, db_session: AsyncSession = Depends(get_db)):
+async def get_user(
+    uuid: UUID,
+    db_session: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user),
+):
     try:
         user = await UserService.get_user(uuid=uuid, db_session=db_session)
 
@@ -94,7 +110,16 @@ async def get_user(uuid: UUID, db_session: AsyncSession = Depends(get_db)):
 
 
 @router.delete("/{uuid}")
-async def delete_user(uuid: UUID, db_session: AsyncSession = Depends(get_db)):
+async def delete_user(
+    uuid: UUID,
+    db_session: AsyncSession = Depends(get_db),
+    current_user: User = Depends(get_current_user),
+):
+    if current_user.uuid != uuid:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Operation not permitted",
+        )
     try:
         await UserService.delete_user(db_session=db_session, uuid=uuid)
